+++ src/lib/Spider.ts
@@ -11,7 +11,7 @@ const SummarizerManager = require("node-summarizer").SummarizerManager
 const cheerio = require('cheerio')
 
 const bunyan = require('bunyan')
-const bformat = require('bunyan-format')  
+const bformat = require('bunyan-format2')  
 const formatOut = bformat({ outputMode: 'short' })
 const log = bunyan.createLogger({src: true, stream: formatOut, name: "spider"})