+++ src/lib/cov.ts
@@ -1,6 +1,6 @@
 
 const bunyan = require('bunyan')
-const bformat = require('bunyan-format')  
+const bformat = require('bunyan-format2')  
 const formatOut = bformat({ outputMode: 'short' })
 const log = bunyan.createLogger({src: true, stream: formatOut, name: "cov"})
 import fs = require('fs-extra')