+++ src/CHANGELOG.md
@@ -1,5 +1,7 @@
 
+- gitdown no longer removes the folder: it just adds.
 - removed LOCALFolder in gitdown, uses REPOfolder
+
 
 ### 1.11.5
 - moved DB sql here from mbake